// src/app/redux-flow/store/VOD/Engagement/types.ts
export enum ActionTypes {
    GET_VOD_ENGAGEMENT_SETTINGS = "@@vod_engagement/GET_VOD_ENGAGEMENT_SETTINGS",
    SAVE_VOD_ENGAGEMENT_SETTINGS = "@@vod_engagement/SAVE_VOD_ENGAGEMENT_SETTINGS",
    SAVE_VOD_AD = "@@vod_engagement/SAVE_VOD_AD",
    CREATE_VOD_AD = "@@vod_engagement/CREATE_VOD_AD",
    DELETE_VOD_AD = "@@vod_engagements/DELETE_VOD_AD",
    GET_UPLOAD_URL= "@@vod_engagements/GET_UPLOAD_URL",
    UPLOAD_IMAGE= "@@vod_engagements/UPLOAD_IMAGE",
    DELETE_IMAGE="@@vod_engagements/DELETE_IMAGE"
}